some model

+++ Trip.JinJiang.H5/Models/LineLists/EsGroup.cs
@@ -12,7 +12,7 @@ namespace Trip.JinJiang.H5
     {
         public int groupId { get; set; }
         public int lineId { get; set; }
-        public int salePrice { get; set; }
+        public decimal salePrice { get; set; }
         public int clusterType { get; set; }
         public object deadLineTime { get; set; }
         public string deadLineTimeVal { get; set; }
